+++ alvr/client_core/src/connection.rs
@@ -38,20 +38,20 @@ use crate::audio;
 use alvr_audio as audio;
 
 const INITIAL_MESSAGE: &str = concat!(
-    "Searching for server...\n",
+    "Searching for streamer...\n",
     "Open ALVR on your PC then click \"Trust\"\n",
     "next to the client entry",
 );
 const NETWORK_UNREACHABLE_MESSAGE: &str = "Cannot connect to the internet";
 // const INCOMPATIBLE_VERSIONS_MESSAGE: &str = concat!(
-//     "Server and client have\n",
+//     "Streamer and client have\n",
 //     "incompatible types.\n",
 //     "Please update either the app\n",
 //     "on the PC or on the headset",
 // );
 const STREAM_STARTING_MESSAGE: &str = "The stream will begin soon\nPlease wait...";
-const SERVER_RESTART_MESSAGE: &str = "The server is restarting\nPlease wait...";
-const SERVER_DISCONNECTED_MESSAGE: &str = "The server has disconnected.";
+const SERVER_RESTART_MESSAGE: &str = "The streamer is restarting\nPlease wait...";
+const SERVER_DISCONNECTED_MESSAGE: &str = "The streamer has disconnected.";
 
 const DISCOVERY_RETRY_PAUSE: Duration = Duration::from_millis(500);
 const RETRY_CONNECT_MIN_INTERVAL: Duration = Duration::from_secs(1);